Move TestNG "dummy" tests to testFixtures
