+++ subprojects/testing-jvm/build.gradle.kts
@@ -60,6 +60,9 @@ dependencies {
         because("Tests instantiate DefaultClassLoaderRegistry which requires a 'gradle-plugins.properties' through DefaultPluginModuleRegistry")
     }
     integTestDistributionRuntimeOnly(project(":distributions-jvm"))
+
+    testFixturesCompileOnly(project(":testing-base"))
+    testFixturesCompileOnly(libs.testng)
 }
 
 strictCompile {
@@ -69,16 +72,6 @@ strictCompile {
 
 packageCycles {
     excludePatterns.add("org/gradle/api/internal/tasks/testing/**")
-}
-
-tasks.test {
-    exclude("org/gradle/api/internal/tasks/testing/junit/AJunit*.*")
-    exclude("org/gradle/api/internal/tasks/testing/junit/BJunit*.*")
-    exclude("org/gradle/api/internal/tasks/testing/junit/ATestClass*.*")
-    exclude("org/gradle/api/internal/tasks/testing/junit/ATestSetUp*.*")
-    exclude("org/gradle/api/internal/tasks/testing/junit/ABroken*TestClass*.*")
-    exclude("org/gradle/api/internal/tasks/testing/junit/ATestSetUpWithBrokenSetUp*.*")
-    exclude("org/gradle/api/internal/tasks/testing/testng/ATestNGFactoryClass*.*")
 }
 
 integTest.usesJavadocCodeSnippets.set(true)
